--- FanControl/conf/init.json
@@ -1,4 +1,0 @@
-{
-  "test": "test",
-  "hello": "hello"
-}
