+++ FanControl/conf/params.json
@@ -1,0 +1,4 @@
+{
+  "os": "none",
+  "language": "en"
+}
