+++ FanControl/src/jvmMain/kotlin/ui/screen/body/fanList/FanViewModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.update
 import model.hardware.Sensor
 import model.item.SensorItem
 import ui.utils.Resources
-import ui.utils.checkNameTaken
+import utils.checkNameTaken
 
 class FanViewModel(
     private val _fanItemList: MutableStateFlow<SnapshotStateList<SensorItem>> = State._fanItemList,
